bashyrc: remove brew-gist-logs, brew-build-* functions
brew-gist-logs is obsolete. It was mostly about GitHub API token use, and I don't do that any more, and they (GH and Homebrew) have probably changed how it works by now.

The brew-build-* functions, I don't want to use those any more. I don't build from source often, and when I do, it's for debugging, so I think it's better to have the standard commands all written out on the command line, so I know exactly what was run, and it's suitable for sharing in bug reports and the like.